requirements.txt
h2o_wave==0.20.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
scikit_image==0.19.1
torch==1.10.2
torchvision==0.11.3
